requirements.txt
Django>=3.2
 djangorestframework>=3.12
 django-cors-headers>=3.7
